Started to split the Disk Cache from Legacy project loader
