+++ extide/gradle/src/org/netbeans/modules/gradle/GradleAuxiliaryPropertiesImpl.java
@@ -19,7 +19,6 @@
 
 package org.netbeans.modules.gradle;
 
-import org.netbeans.modules.gradle.loaders.LegacyProjectLoader;
 import org.netbeans.modules.gradle.spi.GradleFiles;
 import java.io.File;
 import java.io.FileInputStream;
@@ -28,8 +27,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.HashSet;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Set;
 import org.netbeans.api.project.ProjectManager;
 import org.netbeans.spi.project.AuxiliaryProperties;
@@ -125,6 +122,6 @@ public class GradleAuxiliaryPropertiesImpl implements AuxiliaryProperties {
 
     private File getPropFile(boolean shared) {
         GradleFiles gf = project.getGradleFiles();
-        return new File(shared ? gf.getProjectDir() : LegacyProjectLoader.getCacheDir(gf), GradleFiles.GRADLE_PROPERTIES_NAME);
+        return new File(shared ? gf.getProjectDir() : NbGradleProjectImpl.getCacheDir(gf), GradleFiles.GRADLE_PROPERTIES_NAME);
     }
 }
